Fix: 품목명 select로 변경 && 입력 폼 형식 변경
입력 등의 작업을 위해서는 품목명을 알아야할 필요가 있음
input text에서는 품목명을 정확히 입력해야하므로
작업이 힘들기 때문에 select로 변경
입력 폼 형식 변경

+++ src/main/java/com/mit/model/PlanDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class PlanDTO {
 	Long planNum;
-	Integer eplNum;
+	Long eplNum;
 	Long productNum;
 	Long quantity;
 	java.sql.Date dueDate;
@@ -15,6 +15,6 @@ public class PlanDTO {
 	
 	String productName;
 	String partnerName;
-	Integer productPrice;
+	Long productPrice;
 	
 }
